Moved models from MedievalBattle to Database
Moved models && created migrations
