+++ Backend-dotnet/MessengerGamingPlatform/Database/Models/MedievalBattle.cs
@@ -3,6 +3,7 @@ using System.Collections.Generic;
 using System.ComponentModel.DataAnnotations;
 using System.Linq;
 using System.Threading.Tasks;
+using Database.Models.MedievalBattleModels;
 
 namespace Database.Models
 {
@@ -12,6 +13,8 @@ namespace Database.Models
         public int MedievalBatlleId { get; set; }
         public int SessionMedievalBattleId { get; set; }
         public SessionMedievalBattle SessionMedievalBattle { get; set; }
-        public List<Player> Players { get; set; }
+        public AbstractField AbstractField { get; set; }
+
+        // public List<Player> Players { get; set; }
     }
 }
