+++ Backend-dotnet/MessengerGamingPlatform/Database/Models/MedievalBattleModels/AbstractField.cs
@@ -4,7 +4,7 @@ using System.ComponentModel.DataAnnotations;
 using System.Linq;
 using System.Threading.Tasks;
 
-namespace MedievalBattle.Models
+namespace Database.Models.MedievalBattleModels
 {
     public class AbstractField
     {
