+++ Backend-dotnet/MessengerGamingPlatform/Database/Models/MedievalBattleModels/GameController.cs
@@ -1,12 +1,11 @@
-﻿using MedievalBattle.Models.Primitives;
-using System;
+﻿using System;
 using System.Collections.Generic;
 using System.ComponentModel.DataAnnotations;
 using System.Linq;
 using System.Threading.Tasks;
 using Database.Models;
 
-namespace MedievalBattle.Models
+namespace Database.Models.MedievalBattleModels
 {
     public class GameController
     {
