+++ Backend-dotnet/MessengerGamingPlatform/Database/Models/MedievalBattleModels/LocalStatistic.cs
@@ -3,7 +3,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
 
-namespace MedievalBattle.Models
+namespace Database.Models.MedievalBattleModels
 {
     public class LocalStatistic
     {
