+++ Backend-dotnet/MessengerGamingPlatform/Database/Models/MedievalBattleModels/UnitClasses/Archer.cs
@@ -4,9 +4,9 @@ using System.ComponentModel.DataAnnotations;
 using System.Linq;
 using System.Threading.Tasks;
 
-namespace MedievalBattle.Models.UnitClasses
+namespace Database.Models.MedievalBattleModels
 {
-    public class Flank : AbstractField
+    public class Archer : AbstractField
     {
 
     }
