+++ Backend-dotnet/MessengerGamingPlatform/Database/Models/MedievalBattleModels/UnitClasses/Fighter.cs
@@ -4,7 +4,7 @@ using System.ComponentModel.DataAnnotations;
 using System.Linq;
 using System.Threading.Tasks;
 
-namespace MedievalBattle.Models.UnitClasses
+namespace Database.Models.MedievalBattleModels
 {
     public class Fighter : AbstractField
     {
