+++ Backend-dotnet/MessengerGamingPlatform/Database/Models/MedievalBattleModels/UnitClasses/Flank.cs
@@ -4,12 +4,10 @@ using System.ComponentModel.DataAnnotations;
 using System.Linq;
 using System.Threading.Tasks;
 
-namespace MedievalBattle.Models.Primitives
+namespace Database.Models.MedievalBattleModels
 {
-    public class Coin
+    public class Flank : AbstractField
     {
-        [Key]
-        public int coinId { get; set; }
-        public int value { get; set; }
+
     }
 }
