--- Backend-dotnet/MessengerGamingPlatform/MedievalBattle/Models/UnitClasses/Archer.cs
@@ -1,13 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.ComponentModel.DataAnnotations;
-using System.Linq;
-using System.Threading.Tasks;
-
-namespace MedievalBattle.Models.UnitClasses
-{
-    public class Archer : AbstractField
-    {
-
-    }
-}
